+++ socialhome/content/enums.py
@@ -1,5 +1,6 @@
 import warnings
 
+from django.utils.translation import ugettext_lazy as _
 from enumfields import Enum
 
 
@@ -8,3 +9,14 @@ class ContentTarget(Enum):
 
     warnings.warn("ContentTarget has been removed. "
                   "Remove code once migrations have been squashed.", UserWarning)
+
+
+class ContentType(Enum):
+    CONTENT = 0
+    REPLY = 1
+    SHARE = 2
+
+    class Labels:
+        CONTENT = _("Content")
+        REPLY = _("Reply")
+        SHARE = _("Share")