+++ socialhome/federate/utils/tasks.py
@@ -4,6 +4,7 @@ import django_rq
 from federation.entities import base
 from federation.fetchers import retrieve_remote_profile
 
+from socialhome.content.enums import ContentType
 from socialhome.content.models import Content
 from socialhome.content.utils import safe_text, safe_text_for_markdown
 from socialhome.enums import Visibility
@@ -50,6 +51,8 @@ def process_entities(entities):
                 process_entity_follow(entity, profile)
             elif isinstance(entity, base.Profile):
                 Profile.from_remote_profile(entity)
+            elif isinstance(entity, base.Share):
+                process_entity_share(entity, profile)
         except Exception as ex:
             logger.exception("Failed to handle %s: %s", entity.guid, ex)
 
@@ -223,6 +226,41 @@ def process_entity_retraction(entity, profile):
         logger.debug("Ignoring retraction of entity_type %s", entity_type)
 
 
+def process_entity_share(entity, profile):
+    """Process an entity of type Share."""
+    if not entity.entity_type == "Post":
+        # TODO: enable shares of replies too
+        logger.warning("Ignoring share entity type that is not of type Post")
+        return
+    try:
+        target_content = Content.objects.get(guid=entity.target_guid, share_of__isnull=True)
+    except Content.DoesNotExist:
+        # Try fetching
+        # TODO federation library fetch util, for now just return
+        logger.warning("No target found for share even after fetching from remote: %s", entity)
+        return
+    if not target_content.author.handle == entity.target_handle:
+        logger.warning("Share target handle is different from the author of locally known shared content!")
+        return
+    values = {
+        "text": safe_text_for_markdown(entity.raw_content),
+        "author": profile,
+        # TODO: ensure visibility constraints depending on shared content?
+        "visibility": Visibility.PUBLIC if entity.public else Visibility.LIMITED,
+        "remote_created": safe_make_aware(entity.created_at, "UTC"),
+        "service_label": safe_text(entity.provider_display_name) or "",
+    }
+    values["text"] = _embed_entity_images_to_post(entity._children, values["text"])
+    guid = safe_text(entity.guid)
+    content, created = Content.objects.update_or_create(guid=guid, share_of=target_content, defaults=values)
+    if created:
+        logger.info("Saved share: %s", content)
+    else:
+        logger.info("Updated share: %s", content)
+    # TODO: send participation to the share from the author, if local
+    # We probably want that to happen even though our shares are not separate in the stream?
+
+
 def _make_post(content):
     try:
         return base.Post(
@@ -255,7 +293,7 @@ def _make_comment(content):
 def make_federable_content(content):
     """Make Content federable by converting it to a federation entity."""
     logger.info("make_federable_content - Content: %s", content)
-    if content.parent:
+    if content.content_type == ContentType.REPLY:
         return _make_comment(content)
     return _make_post(content)
 
@@ -264,8 +302,11 @@ def make_federable_retraction(content, author):
     """Make Content retraction federable by converting it to a federation entity."""
     logger.info("make_federable_retraction - Content: %s", content)
     try:
+        entity_type = "Comment" if content.content_type == ContentType.REPLY else \
+            "Share" if content.content_type == ContentType.SHARE else "Post"
         return base.Retraction(
-            entity_type="Comment" if content.parent else "Post",
+            # TODO check share federation
+            entity_type=entity_type,
             handle=author.handle,
             target_guid=content.guid,
         )